8.15 User Profile End
populate, relationship

+++ src/controllers/videoController.js
@@ -1,23 +1,9 @@
-import Video, {formatHashtags} from "../model/Video";
-
-/* 
-<----- CallBack fuction ------> 
-console.log("i start"); // 1
-Video.find({}, (error, videos) => {
-    if(error) {
-        return res.render("server-error");
-    }
-    return res.render("home", {pageTitle : "Home", title : "Hello", videos:[]}); // 3
-});
-console.log("i finished"); //2 
-콜백함수는 함수가 실행될때까지 기다리는 동안 다른 실행문이 실행됨
-promise는 함수가 실행될때까지 기다림 (다른 실행문도 기다림)
-*/
+import Video from "../model/Video";
+import User from "../model/User";
 
 export const home = async(req, res) => {
     try {
         const videos =  await Video.find({}).sort({ createdAt : "desc"})
-        // console.log(videos);
         return res.render("home", {pageTitle : "Home", videos});
     }
     catch {
@@ -26,30 +12,38 @@ export const home = async(req, res) => {
     
 } 
 export const getEdit = async (req, res) => {
-    const { id } = req.params; //ES6 version / it's same "const id = req.params.id;"
+    const { id } = req.params;
+    const { user: { _id }} = req.session;
     const video = await Video.findById(id);
     // getEdit에서는 exists를 못쓰고 findById 사용 : video Object가 필요하기 때문
     if (!video) {
         return res.render("404", {pageTitle: "Video is not found"});
     }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     return res.render("edit", {pageTitle: `Editing: ${video.title} `, video})
 }
 export const postEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id }} = req.session;
     const { title, description, hashtags } = req.body;
-    const video = await Video.exists({ _id:id}); 
+    const video = await Video.findById(id);
     // exists(filter): video Object를 받는 대신 True or Flase를 return 받음 
     if (!video) {
         return res.status(404).render("404", {pageTitle: "Video is not found"});
-    }
+    };
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    };
     await Video.findByIdAndUpdate(id, {
         title, description, hashtags: Video.formatHashtags(hashtags)
-    })
+    });
     return res.redirect(`/videos/${id}`);
 }
 export const watch = async (req, res) => {
     const { id } = req.params; // id 는 req.params에서 오는 것 기억하기
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate('owner');
     if(!video) {
         return res.status(404).render("404", {pageTitle: "Video is not found"});
     }
@@ -60,15 +54,20 @@ export const getUpload = (req, res) => {
 }
 
 export const postUpload = async (req, res) => {
-    const { path: fileUrl } = req.file;
-    const { title, description, hashtags } = req.body;
+    const { session: { user: { _id }},
+            file: { path: fileUrl },
+            body: { title, description, hashtags }} = req;
     try{
-        await Video.create({
+        const newVideo = await Video.create({
             title, 
             fileUrl,
             description,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags)
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/")    
     }
     // await ~.create : 데이터를 만들고 저장까지 함
@@ -80,7 +79,18 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
-    await Video.findByIdAndDelete(id) 
+    const { user: { _id }} = req.session;
+    const video = await Video.findById(id);
+    const user = await User.findById(_id);
+    if (!video) {
+        return res.render("404", {pageTitle: "Video is not found"});
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+    user.videos.splice(user.videos.indexOf(id), 1);
+    user.save();
+    await Video.findByIdAndDelete(id);
     // findByIdAndRemove도 있지만 되도록 Delete 사용
     return res.redirect("/");
 }
